build: Added run.sh script to set up and check yarn version using fzf for interactive selection
This commit adds a new run.sh script that checks if the 'fzf' command is available in the system, sets up yarn with corepack and stable version, and allows the user to select the job to execute using fzf. This improves the user experience by providing an interactive way to set up yarn without having to manually run commands.